requirements.txt
Pillow==9.5.0
tk=0.1.0